server/requirements.txt
fastapi
uvicorn
pika
psycopg2-binary
boto3
python-multipart
